HB compatible glyf processing
Adds a new `glyf` module that processes outlines in a way that matches HarfBuzz and calls into it for drawing.

Currently generates bit identical values for all test cases but the rounding in the SVG files is difficult to match. This patch contains a very ugly hack to make the all of the tests pass.